+++ R/docs_interview_manual.R
@@ -29,7 +29,7 @@ interview_manual <- function(year,
 
   ### check inputs
   if (missing(year) || is.null(year)) { error_year_not_declared() }
-  checkmate::assert_numeric(year)
+  checkmate::assert_number(year)
   checkmate::assert_logical(verbose)
 
   # data available for the years: