+++ R/docs_questionnaire.R
@@ -35,7 +35,7 @@ questionnaire <- function(year,
 
   ### check inputs
   if (missing(year) || is.null(year)) { error_year_not_declared() }
-  checkmate::assert_numeric(year)
+  checkmate::assert_number(year)
   checkmate::assert_logical(verbose, null.ok = FALSE)
 
   # data available for the years: