+++ tests/testthat/test_labels_mortality.R
@@ -1,5 +1,3 @@
-context("add_labels_mortality")
-
 # skip tests because they take too much time
 skip_if(Sys.getenv("TEST_ONE") != "")
 testthat::skip_on_cran()